requirements.txt
Pillow
numpy
tensorflow
torch
torchvision
matplotlib
tqdm
keras
pandas
seaborn
scipy
scikit-image